Refactor generator and load file creation in pg_to_switch.py
This commit renames gen_prebuild_newbuild_info_files() to
generator_and_load_files() and moves parts of it into separate
functions to make the flow of data clearer. It also splits the
main loop (across all years) into separate sections for each group
of files, with separate blocks within each for the clustered time
(sampled) cases and single-timeseries cases. It produces the same
output as the previous version, but should be easier to revise in
the future.