Retrieve unit-level data from gc.all_units instead of gc.units_model
This makes a more stable link to generator data from
gc.create_all_generators() because all_units has enough data to
construct a Resource ID, so we can link directly back on that. We
also can use the retirement_year already stored in all_units
instead of re-creating that for units_model.

A couple of side-effects:

- we now generally show all capacity ever built in gen_build_predetermined
  instead of just the capacity active for this model year. This shouldn't
  cause problems with Switch and may improve consistency between steps
  in myopic models.
- we now use average values across all model years for foresight models
  instead of values from the first year (for derated capacity, fixed O&M,
  etc.)
- there seem to be minor changes in the derated capacity for small hydro
  geothermal and biomass plants
- we no longer have a problem with generators appearing in
  create_all_generators() without matching unit info, except for
  distributed_generation