Derate capacity via gen_forced_outage_rate instead of build_gen_predetermined
This more accurately reflects what the derating shows (energy limitations
for some renewable technologies that would otherwise be able to run 24/7)
and avoids problems with cluster sizes changing between model periods and
inaccurate reporting of total capacity after the model finishes. This will
also help with retirements, making fixed O&M costs more accurate and
avoiding the risk of cases where we retire capacity in one period, but then
the cluster size changes in the next period so we have the wrong amount
available.

(GenX deals with those issues by carrying an `unmodified_cap_size` through;
Switch doesn't have an equivalent of that.)

This commit also includes a change to the way generators are assigned to
balancing areas. It now takes the last balancing area assigned to the
generator instead of the last balancing area assigned for this
generator/balancing area combination (which seems to be a bug). Without
this change, the code was breaking when we moved from PUDL 0.6.0 to
PUDL 2022.11.30, because that introduced some early generator rows
without balancing areas, which were then becoming the final balancing
areas for the generators.