Improve construction schedule inference for generic (distributed) gens
For distributed generation, PowerGenome only reports total capacity
online in each model year, not dates when it was built. We previously
inferred a construction schedule by taking the simple diff in capacity
between subsequent periods. However, that produces a negative
construction plan when capacity decreases (maybe due to older systems
aging out, maybe just an error). The new code sets this up as a
least-squares problem: minimize (Ax - b)^2, subject to x >= 0, where A
is the build_year : model_year correspondence matrix (1 for any build
years that are active in a particular model year), x is the capacity
built each year and b is the capacity online in each model year.

For monotonically increasing capacity or capacity with one dip, this
should always have an exact solution. For more complex patterns,
especially with long retirement ages, an exact solution may not be
possible, in which case a warning will be given.